re-organize test specs
