+++ fortigate-autoscale/aws/test-helper/index.ts
